+++ src/main/java/roomescape/reservation/service/ReservationService.java
@@ -1,10 +1,10 @@
-package roomescape.service;
+package roomescape.reservation.service;
 
 import java.util.List;
 import org.springframework.stereotype.Service;
-import roomescape.controller.request.ReservationRequest;
-import roomescape.controller.response.ReservationResponse;
-import roomescape.repository.ReservationRepository;
+import roomescape.reservation.controller.request.ReservationRequest;
+import roomescape.reservation.controller.response.ReservationResponse;
+import roomescape.reservation.repository.ReservationRepository;
 
 @Service
 public class ReservationService {
